Place Spire build outputs in stable locations.
The build outputs from Spire are currently all placed in locations under `Source/`, and the exact path depends on whether you select the x64 or Win32 target (this is just one of the things Visual Studio defaults in a weird way, for legacy reasons).

This change tries to direct all outputs to directories directly under the project root:

- `bin/` is the output location for the `.exe` projects
- `lib/` is the output location for the `.lib` projects
- `intermediate/` is used as the intermediate directory for compilation, making it easier to inspect or blow away manually

The project currently doesn't have an `include/` directory for the headers that comprise the public interface, but that would likely be a more invasive change, so I'm not attempting it right now.