Add a "hello world" Spire+D3D11 example.
The example is about as minimal as I can make it right now on the C++ side.
There are still some `TODO` comments in places where maybe we can tweak the API to make things simpler (and a few places where there are flat-out missing APIs).
Also, I haven't implemented the de-initialization logic to properly clean up after D3D.

The shader side could use a cleanup pass. In particular, it would be great to revise it to use a truly minimal pipeline declaration, rather than one I copy-pasted from other examples.

As is to be expected, most of the code is just the Windows and D3D11 cruft, but I've tried to keep that as unobtrustive as possible, and move the Spire-related stuff to the front.