Add an API entry point to query name of a `SpireModule`
This can be seen as a workaround for not being able to specify a `SpireModule*` to `spShaderAddModule`, but it also seems like a reasonable convenience function to have anyway.

This change was motivated by work I'm doing to port a small D3D11 testbed over to the new official Spire C API (instead of my hacked up home-made version).