Remove some dead code from CoreLib.
There are entire files in `CoreLib` that aren't being used.
This change deletes them as the first step toward making the codebase as simple as possible.

I fully realize that we might need some of this functionality again some day, but when/if that day comes we can easily recover the code from the Git history.