Remove `CodePack` and `LibraryRelease/`
The amalgamated source file in `LibraryRelease/Spire.cpp` is a nice idea, but it has a lot of practical down-sides:

- It complicates the process for contributing to the code (an extra step before submitting changes, more complicated diffs)
- It adds more code to the repository (`CodePack`, which is in C# rather than C++)
- It creates the potential for confusion over where to make edits and/or can lead to stale changes or hard-to-integrate patches (if people edit the amalgamated file directly).

Some of these down-sides could be mitigated, but it seems a lot simpler to just the `SpireAllSource.h` file be the one-stop-shop for anybody who wants a simple way to pull in Spire without a lot of build complexity.

Once `LibraryRelease/` gets blown away, there is no reason to keep `CodePack` in the repository.