Refactor API for capturing diagnostic output.
Rather than diagnostics only be queryable from a `SpireCompilationResult`, this change adds a distinct `SpireDiagnosticSink` that can be passed to operations that might produce diagnostics.
The client can create multiple sinks, and can query the diagnostics written to a given sink, or clear the sink of output.

I'm not 100% happy with this design, since with the way things currently work, your `SpireCompilationContext` is pretty much hosed as soon as a single operation fails, so we might as well just store the diagnostic info on the context itself (and indeed, the context *does* track an error status). This API is more aspirational, for a point where the compilation context is more robust to failure of individual operations.