Centralize diagnostic message info.
Previously the compiler would generate diagnostic messages in an ad hoc fashion at the point where an error is detected. This has several down-sides:

- Difficult to manage the space of diagnostic IDs, to avoid clashes
- Difficult to ensure that when the same diagnostic is output from multiple places, the messages stay in sync
- Difficult to audit all messages to make changes to terminology, style, etc.
- Difficult to change how types, declarations, etc. get printed (requires ad hoc routines that generate strings or write to a `StringBuilder`
- Would be difficult to localize message text, if this ever became a priority

The new approach requires that diagnostics are all defined centrally in `DiagnosticDefs.h` as follows:

    // made-up error diagnostic
    DIAGNOSTIC(12345, Error, typeIsBad, "the type '$0' is bad")

This declaration includes the numeric message ID, the severity (Error/Warning/etc.), the symbolic name (used to reference this message in code), and a format string for the default (non-localized) message, with `$` used to splice in additional arguments.

A use site that wants to emit a diagnostic writes:

    DiagnosticSink* sink = ...;
    sink->diagnose(someLocation, Diagnostics::typeIsBad, someType);

The first argument is the location (or an argument of a type`T` that supports a `getDiagnosticLocation(T)` function), the second is the diagnostic defined above, and any additional arguments are passed after that. The additional arguments must be of some type `T` that supports `printDiagnosticArg(StringBuilder&, T)`.

No attempt is made to statically check that a call site passes the number and/or type of arguments expected by a diagnostic. Programmers should make sure to write a test case that triggers any diagnostic message they add, to confirm that it works.