Continue work on preprocessor.
This change adds basic support for function-like macros, as well as token pasting (infix `##` operator).

The way that expansion for arguments to function-like macros is handled is a bit non-traditional.
The letter of the spec is that arguments get expanded *before* substitution (but when used with the stringize operator they don't get expanded...), and then everything gets expanded again after substitution.
I'm instead treating macro arguments as a special flavor of macro themselves, and then treating expansions as having an associated "environment" that they use for looking up macros (including both arguments and globally-defined macros).

The environment-based approach to expansion means that there isn't currently much special-case logic for handling arguments, and should make it easy(-ish) to get stringizing correct if we ever need it. The two big down-sides are:

- It complicates the defintion of a macro being "busy"; I think this is tractable to solve, but not a high priority, since "busy"-ness only affects recursive macro invocations.

- It complicates the rules for parsing macro invocations. Right now I read "raw" tokens for a macro invocation (because you don't want to macro-expand when parsing the argument list), but really I need to perform macro expansion, but only on macros that are function arguments. This is pretty subtle, but I'll need to get to it eventually.

One remaining complication for function-like macros is that they create a case where we need two tokens of lookahead to recognize whether macro expansion is needed. If we see an identifier as our lookahead, and that identifier names a function-like macro, we don't know if expansion is really needed until we look at the token *after* the identifier. This is obviously fine if we know we need to consume the identifier anyway (consume it, then peek), but we currently have a bunch of cases where we want to peek at tokens even in the context of macro expansion.

The token pasting `##` operator is handled in a pretty simple fashion. Right after consuming a token (including macro expansion) we check if the next token is `##`, and if so we begin appending tokens to a `StringBuilder` for as long as we see a sequence of tokens separated by `##`. Once we are done collecting tokens, we invoke the lexer to tokenize the string (it is expected to yield a single token, but we don't require this), and push the resulting tokens onto our stack of input.

Token pasting doesn't interact properly with "peeking" tokens either (since we'd need to look ahead two tokens), so this may be a sign that the implementation needs to change to not use peeking so heavily. The main reason why this has been important, though, is that we need to be careful not to move the input cursor past the end of a stream when reading tokens of a directive at the end of that stream (e.g., we don't want to move past the end of a `#include`d file when reading a final `#endif` in that file!).

More work is still needed.