Add a new interface for token reading.
Rather than having clients of the lexer or preprocessor work directly with a `List<Token>` and an `int` position, this change adds a customized `TokenList` and `TokenReader` that help make it easier to do parsing without as many corner cases.

The key ideas are:

- A `TokenList` always includes a token "past the end" representing the end of file (or similar)
- A `TokenReader` will refuse to advance past the end token of its corresponding sequence

This means that it is *always* safe to ask a `TokenReader` about its "current" token, and it is always safe to "read" a token (even though this may return the same end-of-file token over and over).

Right now the end-of-file token doesn't have a proper line number on it (the `Tokenizer` will need to be updated to produce this), so error reporting still won't be perfect. I tweaked the preprocessor a bit, though, to ensure that it returns the EOF token from the main input file it was given, instead of its dummy EOF token (which doesn't have a file name).

The way this is architected right now allows a `TokenReader` to be formed over a sub-sequence of tokens (e.g., all the tokens between a `(` and matching `)`, and it will return a dummy end-of-file token with the correct location information of the `)`. This degree of freedom isn't being used right now, and probably needs to be more carefully thought about (I've used this kind of capability in other parsers I've written).

Switching the `Parser` over to the new interface actually got rid of an internal compiler error (out-of-range indexing of the token sequence) in one of our test cases, so it is already a small improvement.

I also did a small refactoring to the `Parser` code, where I moved the class declaration over to the .cpp, so that the .h just declares the one function anybody uses. This is really to pave the way for me to do more major refactoring of the parser.