Remove `StructSymbol`
We just use `StructSyntaxNode` for everything front-end related, and then have the code generator maintain a map from `StructSyntaxNode*` to `ILStructType` that it fills in as needed.

The main complication in this change is that the SpireLib interface was using IL types to get layout information to provide through the API, but I had already refactored the layout stuff to be a separate interface, so rewriting the code to just use front-end types was easy enough.