Move type layout rules to separate file.
This makes it easier to compute layout consistently across different type representations (e.g., both `ILType` and  `ExpressionType`), becaues the code is centralized.

This may have introduced some behavior diffs if I implemented the rules wrong, but looking at the original code, this seems to require a pass to scrutinize the rules anyway.