Introduce minor sharing between syntax for function-like decalrations.
With this change, both `FunctionSyntaxNode` and `ImportOperatorDefSyntaxNode` inherit from `FunctionDeclBase`, which holds their common members.

As with earlier changes of this kind, the main changes had to do with `FunctionSyntaxNode` now having a `Token` instead of a `String` for its `Name`. (Aside: I'm not sure that having a `Token` for the name of a `Decl` is really a good idea, so it might be worth revisiting why I needed to go that direction in the first place.