Fix function-like macros for `TokenList` change.
The `TokenList` change means that we now expect all token sequences to have an extra token "past the end" that represents the end of the stream. This is helpful for reporting good diagnostics.

When I made the change, I neglected to account for the way that we have arguments to function-like macros behave as if they were macros themselves (so that they hold token lists), and we didn't have any tests covering that case. This change puts in a fix.