Basic `typedef` declaration support.
The easy part of this change was adding the `typedef` syntax (parses to a `TypeDefDecl`).

The more interesting part is how this interacts with the representation of types during semantics (`ExpressionType`).
For good error messages, you want to refer to a `typedef` type by its name, not the definition, so we create a `NamedTypeExpression` to represent it.

Code that wants to check if a type matches a certain case already does, e.g., `type->AsArray()` so it was relatively simple to make these queries deal with the possibility that a named type might be a `typedef` of an array type.

I refactored this part of the `ExpressionType` code a bit to streamline things a bit, but the overall semantics shouldn't have changed.