Parser error recovery without exceptions.
This change alters the way that error recovery is implemented in the main Spire parser.
Previously the parser had thrown exceptions when it encountered an unexpected token, and then caught those exceptions in various places to try to implement recovery.

This change implements an alternative strategy where the parser is either in its normal state, or a "recovery" state.
if an unexpected token is discovered when the parser is in the normal state, an error is reported and we enter the recovery state.

During the recovery state, we do *not* emit additional error messages for unexpected tokens.
If we are looking for a specific token when in the recovery state, we will skip tokens (trying to respect brace matching, and trying not to skip closing tokens) looking for a match.
Whenever a match is found in recovery mode, we exit recovery mode and try to parse as normal again.

This strategy probably still needs some tweaking, to improve its behavior in practice.

As part of this change, I also added some convenience functions for the parser that I find help make things more concise and robust in practice:

- `if(AdvanceToken(tt)) {...}` can be used in place of `if(LookAheadToken(tt)) { ReadToken(tt); ...}`
- `while(!AdvanceIfMatch(tt))` can be used in cases where we want to loop until we find a matching token - it helps to make sure we exit the loop if we reach end-of-file

WIP on parser error recovery

continuing work on parser error recovery

complete improved parser error recovery