Unify representation of "modifiers" in the AST.
Rather than have each `Decl` case have its own custom `bool` members for modifiers, this change adds a single `Modifiers` type that all declarations use to hold all of:

- Simple keyword modifiers like `public` and `in`
- GLSL-style `layout` modifiers
- HLSL-style `[Attributes]`

The simple keywords are stored as bit flags, while the more complex cases are stored as a linked list of heap-allocated objects.

The basic idea here is that the parser allows any kind of modifiers (in any order) on any kind of declaration, and it is up to later semantic analysis passes to enforce any desired constraints.
This change doesn't actually add that semantic analysis enforcement, so it technically allows most modifiers to be used in inappropriate places, where they will be ignored.

The main reason to do this is that we can reduce the amount of lookahead required for disambiguation during declaration parsing, because we can always consume an arbitrarily long prefix of modifiers, and *then* decide what kind of declaration to parse.

This change does not take advantage of that freedom, but it sets us up to exploit it.