Unify parsing of declarations.
There is now a single `ParseDecl` entry point that parses a single declaration of any of the kinds we support, and a single `ParseDeclBody` entry point that is used to parse a sequence of declarations in a context where only declarations (and not statements) are allowed.

All the obvious "container" declaration cases (`struct`, `shader`, `module`, and `pipeline`) now use `ParseDeclBody` and therefore are allowed by the parser to contain any number of declarations of any flavor. Later semantic checks will be needed to enforce the rules we want.

Depending on context, we may want to represent what looks syntactically like a variable declaration using one of a few classes in the AST. To that end, we pass the "container" declaration down into `ParseDecl`, and in the variable case we use it to pick the right kind of declaration to construct.

This detail might get removed if we decide to be more uniform and represent, e.g., both struct fields and local variables using the same declaration kind. Almost certainly, if we add support for member functions in `struct` types, we should try to share as much code as possible with the existing `FunctionSyntaxNode`.

The one glaring non-orthogonality here is that component declarations don't play nice with the above, for a few reasons:

1. A component declaration may represent either a logical variable (with optional initializer) or a logical function (with parameters and an optional body), and the AST doesn't work hard to differentiate the two.

2. Because of (1), the `ComponentSyntaxNode` class doesn't inherit from either `VarDeclBase` or `FunctionDeclBase`, so it can't be used polymorphically in cases where those are expected.

3. Component declarations support syntax sugar for block-structured initializers that ordinary variables don't get.

For (3) observe that the component declaration:

    float foo { ...; return bar; }

is equivalent to the following desugared declarations:

    float foo_() { ...; return bar; }
    float foo = foo_();

Moving forward, we'll need to figure out how to split `ComponentSyntaxNode` into two different classes, to handle the function-like and variable-like cases respectively.