Support for parsing HLSL semantics and `cbuffer` declarations.
Now that we no longer allow a `:` after a declaration to be used to indicate an alternate name for an overloaded component, we can use it as it appears in HLSL to add "semantics" to a declaration.
Semantics are stored as `Modifier`s, just like any decorations that appear before a declaration.

The parsing of buffer declarations is currently straightforward.

This change just implements parsing for these constructs, so no checking occurs, and they don't impact code generation.
Also, the way that modifiers are being handled currently means that semantics get parsed, but then end up getting clobbered by other modifiers associated with a declaration.