Add dedicated vector and matrix types to front-end.
Rather than represent vector and matrix types just as certain cases of the `BaseType` enumeration, we instead have dedicated types for `VectorExpressionType` and `MatrixExpressionType` that can conceivably represent any vector/matrix we come up with.
In particular, this should mean that the front-end has no problems supporting non-square matrices.

This also adds support for the HLSL `vector<T,N>` and `matrix<T,R,C>` syntaxes, which are used to define all the usual HLSL and GLSL type names as typedefs.
These types are declared in the stdlib, so that they can be found by ordinary name lookup, and they therefore use some new declaration syntax.

First, there is a syntax for "magic" type declarations that the compiler needs to know about:

    __magic_type(Foo) Bar; // compiler associated this with the name "Foo"

and then there is a helper declaration type `GenericDecl` for generics:

    __generic<T /*type parameter*/, let n : int /*value parameter*/ >  /* any other declaration goes here */

All of this is implemented in a somewhat rudimentary fashion, of course, but it represents progress.

WIP: Trying to remove builtin vector types and add VectorExpressionType

more WIP on generic vector type

Adding dedicated vector and matrix types to front-end.