Remove l-value status from `BasicType`.
Instead, the l-value status is now stored in a struct called `QualType` that is used for the type of an expression.
The main up-sides to this change are:

- The compiler can handle l-values of any type uniformly (we don't have to wedge everything under `BasicType`)
- There should no longer be any reason to clone or modify an `ExpressionType` because the only interesting status is stored separately.

This change of course bloats the amount of storage used to store types on `ExpressionSyntaxNode`.
A possible future optimization would be to store the l-value status as a low-order bit in the same word as the `ExpressionType` pointer.