Work toward generalizing `VarExpressionSyntaxNode`
I changed the lookup logic to try to enforce that if the ordinary `Scope`-based lookup finds something, then *that* is what the expression should use, and it will try to handle all those cases more directly.
I expect that there will still be cases where the older lookup approach matters for quite some time, but this is a start.

Note: this needs a lot more testing, but since I'm working in my fork at the moment, I'm going to blaze ahead.