Get rid of `TypeSyntaxNode`.
Previously, there were some cases where `TypeSyntaxNode` was used for types, and other cases where a type would slip in as an `ExpressionSyntaxNode` simply because there was no way for the parser to disambiguate.

This change makes things uniform, so that we always use `ExpressionSyntaxNode` to represent terms in the AST, even when we expect those terms to evaluate to types.

Details:

- Looking up the names of builtin types is no longer special-cased. They are now declared (with a bit of special-case syntax) in the stdlib, and can be looked up using the ordinary rules.

- Expressions that evaluate to a type now have a special type applied: `TypeExpressionType`, which represents "the type of a type." This means that eventually the `typeResult` contrivance in the semantics visitor might not be needed.

- An array-indexing expression like `A[3]` might represent an actual indexing operation or the description of an array type, and so we don't differentiate the two in the syntax.

This hasn't been fully tested and probably breaks things, but it represents important forward progress, so I'm going to keep working.