Add distinct `FuncType` case for types.
This were previously jammed into `BasicExpressionType`.
Moving them out is part of a move toward having the type hierarchy be more rationalized, to make it possible to add more interesting flavors of builtin/known types.

The one big assumption here is that the `Component` field in `BasicExpressionType` is only ever being used in the case where we are constructing a type to reference a component *function*, and it is unused for ordinary component references.
That needs to be double-checked at some point.