Built-in types are now declared as if they were ordinary `struct` types.
This may actually break some things in back-end code generation, so I'll need to test this change before any pull request.

The basic idea is that now the built-in types get declared directly in the stdlib, and just use a special modifier to indicate their status to the rest of the compiler:

    __builtin_type(32) struct float {};

The argument to `__builtin_type` is the integer value of the corresponding `BaseType` enumeration.

When constructing an `ExpressionType` to reference a particular `Decl`, we check whether it has this modifier, or the similar `__magic_type` modifiers, and use this to select a different concrete subclass of `DeclRefType` to carry any extra semantic baggage.