Turn builtin texture types into generics.
This required adding some very basic support for "default arguments" on generics, since the type `Texture2D` is actually shorthand for `Texture2D<float4>`.
The alternative approach would have been to allow for overloading of type names, so that you'd have both the generic type and then a non-generic `typedef` with the same name, but adding really robust overloading support to Spire seems to be on the "future work" pile.