Work on support for member functions.
With this change, the `Texture2D<T>` type defines a member function `Load` in the standard library, so that we can resolve calls to that function with ordinary rules for member lookup and then function application.
The down-side is that this tears up a bunch of other infrastructure, so I'm pretty sure everything else is completely broken.

In order to support this change, I also added a bit of work towards allowing lazy/recursive declaration checking.
The basic idea is that every `Decl` tracks a flag for its state (unchecked, checking, checked), and we can call a function to "ensure" that a declaration is ready to be used at some use site.
This is a simplistic way to achieve out-of-order declarations, although it has the down-side of requiring unbounded recursion in semantic checking.