Add rudimentary support for initializers and extensions.
(Note: I'm using double underscores for things that are currently only needed in the stdlib. Some of these might be good to expose to end users eventually, but for right now I don't want to think about pretty syntax, so I'm keeping things ugly on purpose.)

Initializers/constructors are written with an `__init` keyword, and are basically just function declarations that are expected to return the type of the enclosing type. I haven't yet worried about what the body of these should look like, because all of the ones I've declared so far should be built-in.

An example initializer would be:

    struct MyThing {
        __init(int x);
    }

    // now I can do:
    MyThing thing = MyThing(99);

The main reason to favor initializers/constructors over the current practice of defining functions with the same names as types is that initializers will work even with `typedef`s. That is, we can do:

    typedef MyThing Foo;
    Foo foo = Foo(99); // works just as well

The other feature is "extension" declarations (keyword: `__extension`), which are akin to `impl` blocks in Rust, or `extension` declarations in Swift. The basic idea with these is that they let us add additional declarations to specific types. The main reason this is helpful/needed is that once we declare the `vector` type as a single generic:

    __generic<T, val N : int> __builtin_type(Vector) struct vector {};

then we can't easily define constructors in it where the signature might depend on `N`. The way around this is to declare an extension that only applies to vectors of specific size:

    __generic<T> __extension vector<T,3> {
        __init(T x, T y, T z);
    }

The challenge of going down this route is that you need a way to decide if a particular extension applies to a given type.

The basic approach right now is:

- An extension must apply to a nominal aggergate type, so that we can resolve the type it applies to when checking the extension generically (not for specific parameters)

- We build a linked list of extensions that potentially apply to a given `AggTypeDecl` so that we can enumerate them easily.

- When it comes time to look up things (currently only doing it for initializers), we need to iterate over the extensions that might apply to a given type, and then determine if they *do* apply.

Determining if a generic extension applies to a type requires a constraint solver / unification strategy. This change punts on that bit.