Add a bit more real overload resolution framework
The basic idea is that name lookup can return an "overload set" as a result (rather than just a single declaration), and then at appropriate points we coerce that result to a single value as needed.

When an overload set is used as the base expression for a function call, we can enumerate the various declarations it comprises, and consider each of them as a candidate.
Rather than immediately dismiss candidates that don't work, we track the best one so far (where, e.g., a candidate with the right number of parameters beats out one with the wrong number) - if we find a single best candidate (including the special case where there was just one candidate) we can produce good error messages specific to that candidate, rather than a general "no applicable function" message.

The approach right now is very rudimentary, so please take everything with a grain of salt.