Add a notion of `DeclRef`s
Once you add things like generics to a language, you have the problem that you can't just read information off the AST and expect it to be valid. E.g.:

    struct Foo<T> { T Load(); }

    Foo<int> f;

    int b = f.Load(); // result must be `int` not `T`

At the site where we use a member (the call to `Load()`) we need to somehow propagate information from the declaration site for `f` through to the return type declared for `Load()` inside `Foo`.

A `DeclRef` captures this information. It consists of a pointer to a declaration, plus an optional list of `Substitutions` to apply. Currently there would be one substitution for each generic declaration that lexically surrounds a the target declaration (so only one for `Load()`), and each substitution holds a mapping from generic parameters to their values for the given generic.

In the example above, the type of `f` is a `DeclRefType` where the enclosed `DeclRef` points to `Foo` and a substition list that maps `T` to `int`.
When we look up the `Load` member in `f`, we are doing lookup inside that `DeclRefType` so we can propagate its substitutions over when we construct a `DeclRef` that points to `Load`.
Finally, when we call `Load()` we can use the declaration part of the `DeclRef` to find the result type as written, and then apply the substitions to it.

In order to make it convenient to work with `DeclRef`s (and less likely to accidentally use information taken from a raw `Decl*`), there is a hierarchy of subtypes of `DeclRef` that can be used for handling different types of declarations, and a bit of a "dynamic cast" mechanism for checking the type of declaration being referenced.

Aside: all of this machinery is currently being used for generics, but it may also be useful for other purposes depending on how advanced the type system for shader components needs to be.
It would note that in Spark, I needed quite similar machinery to get the types of records/rates to work out properly in the context of inheritance, because a method in `Base` that returns `Base::CoarseVertex` needs to actually be considered to return `Derived::CoarseVertex` when called in `Derived`.