Rudimentary scoping for `cbuffer` declarations.
This is a bit messy (I hope to clean it up), but the basic idea is two things:

1. Certain declarations can be marked as "transparent", which means that lookup in the scope of the parent declaration should "see through" them and find members inside.

2. During name lookup, the results include both a declaration reference and a (possibly empty) chain of "breadcrumbs" that represent transparent declarations that were stepped through on the way to the result. If/when we actually create an expression to represent the declaration, we need to walk the breadcrumb trail and generate the appropriate `MemberExpressionSyntaxNode`s along the way.

At this point I can pass a (simple) raw HLSL shader into the front-end and have it get past semantic analysis (I can't guarantee that it is fully and correctly type-checked), before erroring out inside code generation.
The issue in the code generator is that it does not support global variables, and doesn't currently have an obvious representation for the types of constant buffers.