Bug fix for `struct` test case.
The root cause here was that the semantics checking for project expressions was visiting the base expression, but ignoring the result.
The way that name lookup is being handled right now means that the result of checking an expression (or type) could be a different object, so it isn't safe to assume to call the semantics visitor and ignore the result.