Remove legacy notion of "generic types"
Now all generic types are just declared in the stdlib, and use an ordinary generic parameter list.

I went ahead and disabled a bunch of code related to special handling for some of these types, so I'm sure some things are broken, but as with most of the recent changes, the hope is that we can reconstruct the same functionality in a cleaner fashion with the new foundations.