Change type setup for constant buffers to match D3D12 HLSL.
In "shader model 5.1" the new way to declare a constant buffer is with a generic type `ConstantBuffer<T>`:

    struct Foo { int a; float b; };
    ConstantBuffer<Foo> myConstantBuffer : register(b0);

The old `cbuffer` syntax is kept around, but it is just sugar for the above (plus the magic scoping rules).

We were already following this basic approach in Spire, by splitting a `cbuffer` declaration into a type plus a variable declared using that type, but the decomposition I used was treating a constant-buffer type as a special kind of aggregate type, rather than just a generic wrapper around a user-defined `struct` type. This change make the desugaring mirror the above HLSL code.

In order to support this style, I had to add a notion of "pointer-like" types with implicit dereferencing.
That is, when you write like:

    float x = myConstantBuffer.b;

there is a bit of sleight-of-hand going on, because you are treating `myConstantBuffer` like it has type `Foo`, when its real type is `ConstantBuffer<Foo>`.
This is basically just like how many modern langauges treat a pointer to `Foo` just like `Foo` in the context of a member lookup operation.

So, we add a notion of `PointerLikeType` to the type hierarchy, and make `ConstantBuffer<T>` be just such a type.
When we are doing member lookup (*including* member lookup through transparent declarations) we automatically deference any pointer-like types and look at the members of the pointer-to type.
The implicit dereferencing construct explicit `DerefExpr` nodes in the AST, so that downstream phases of compilation don't need to deal with this sublety.